pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
    "setuptools",
    "setuptools-scm",
]

[project]
authors = [
    {email = "temp@gmail.com", name = "UCL ARC"},
]
classifiers = [
    "Operating System :: POSIX",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Typing :: Typed",
]
dependencies = [
    "github3-py==4.0.1",
    "gitpython==3.1.36",
    "keyring==24.2.0",
    "loguru==0.7.2",
    "typer==0.9.0",
]
description = "Quality Workflow management helper. Automating traceability matrices"
dynamic = [
    "version",
]
keywords = [
]
name = "qw"
optional-dependencies = {dev = [
    "black",
    "build",
    "mypy",
    "pre-commit",
    "pytest",
    "ruff",
    "tox>=4",
    "twine",
]}
readme = "README.md"
requires-python = ">=3.11"
license.file = "LICENCE.md"

[project.scripts]
qw = "qw.cli:app"

[project.urls]
homepage = "https://github.com/UCL-ARC/qw"

[tool.coverage]
report = {skip_covered = true, sort = "cover"}
run = {branch = true, parallel = true, source = [
    "qw",
]}
paths.source = [
    "src",
    ".tox*/*/lib/python*/site-packages",
]

[tool.mypy]
explicit_package_bases = true
plugins = [
    "numpy.typing.mypy_plugin",
    "pydantic.mypy",
]

[tool.poetry]
authors = ["Stef Piatek", "Tim Band"]
description = "Tool to help regulatory compliance for projects hosted on github"
name = "qw"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
github3-py = "^4.0.1"
gitpython = "^3.1.36"
keyring = "^24.2.0"
loguru = "^0.7.2"
python = "^3.8"
typer = "^0.9.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
build = "^1.0.3"
mypy = "^1.5.1"
pre-commit = "^3.4.0"
ruff = "^0.0.290"
tox = "^4.11.3"
twine = "^4.0.2"

[tool.pytest.ini_options]
addopts = "--color=yes -v"
testpaths = [
    "tests",
]

[tool.ruff]
fix = true
force-exclude = true
ignore = [
    "D203", # no-blank-line-before-class
    "D212", # multi-line-summary-first-line
    "D417", # argument description in docstring (unreliable)
]
per-file-ignores = {"test_*" = [
    "S101",
]}
select = [
    "A",
    "ARG",
    "B",
    "BLE",
    "C",
    "COM",
    "D",
    "DJ",
    "DTZ",
    "E",
    "EM",
    "ERA",
    "EXE",
    "F",
    "FBT",
    "I",
    "ICN",
    "ISC",
    "N",
    "NPY",
    "PD",
    "PGH",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "PYI",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "S",
    "SIM",
    "T",
    "TCH",
    "TID",
    "UP",
    "W",
    "YTT",
]
target-version = "py311"
isort.known-first-party = [
    "qw",
]
mccabe.max-complexity = 18
pep8-naming.classmethod-decorators = [
    "classmethod",
    "pydantic.validator",
]

[tool.setuptools_scm]
local_scheme = "no-local-version"
write_to = "src/qw/_version.py"

[tool.tomlsort]
all = true
spaces_indent_inline_array = 4
trailing_comma_inline_array = true
overrides."project.classifiers".inline_arrays = false
overrides."tool.coverage.paths.source".inline_arrays = false

[tool.tox]
legacy_tox_ini = """
    [gh-actions]
    python =
        3.11: py311

    [testenv]
    commands =
        pytest --cov --cov-report=xml
    deps =
        pytest
        pytest-cov

    [tox]
    env_list =
        py311
"""
